login,register,authentication added

+++ app/Http/Controllers/ListingController.php
@@ -10,86 +10,109 @@ use Illuminate\Validation\Rule;
 class ListingController extends Controller
 {
     //show all listings
-    public function index(){
+    public function index()
+    {
 
         return view('listings.index', [
             'heading' => 'Latest Listings',
-            'listings' => Listing::latest()->filter(request(['tag','search']))->paginate(6)
+            'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
         ]);
     }
 
     //show single listing
-    public function show(Listing $listing){
+    public function show(Listing $listing)
+    {
         return view('listings.show', [
             'listing' => $listing
         ]);
     }
 
     //show create form
-    public function create(){
+    public function create()
+    {
         return view('listings.create');
     }
 
     //store listing data
-    public function store(Request $request){
+    public function store(Request $request)
+    {
         $formFields = $request->validate([
             'title' => 'required',
-            'company' => ['required',Rule::unique('listings','company')],
+            'company' => ['required', Rule::unique('listings', 'company')],
             'location' => 'required',
             'website' => 'required',
-            'email' => ['required','email'],
+            'email' => ['required', 'email'],
             'tags' => 'required',
             'description' => 'required'
         ]);
 
-        if($request->hasFile('logo')){
-            $formFields['logo'] = $request->file('logo')->store('logos','public');
+        if ($request->hasFile('logo')) {
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
         // Session::flash('message','Listing Created..!');
 
-        return redirect('/')->with('message','Listing created successfully..!');
-
+        return redirect('/')->with('message', 'Listing created successfully..!');
     }
 
     //show edit form
-    public function edit(Listing $listing){
+    public function edit(Listing $listing)
+    {
 
-        return view('listings.edit',['listing' => $listing]);
+        return view('listings.edit', ['listing' => $listing]);
     }
 
 
     //update listing data
-    public function update(Request $request,Listing $listing){
+    public function update(Request $request, Listing $listing)
+    {
+
+        //make sure logged in user is owner
+        if ($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized Action..!');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
             'location' => 'required',
             'website' => 'required',
-            'email' => ['required','email'],
+            'email' => ['required', 'email'],
             'tags' => 'required',
             'description' => 'required'
         ]);
 
-        if($request->hasFile('logo')){
-            $formFields['logo'] = $request->file('logo')->store('logos','public');
+        if ($request->hasFile('logo')) {
+            $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
         $listing->update($formFields);
 
         // Session::flash('message','Listing Created..!');
 
-        return back()->with('message','Listing updated successfully..!');
-
+        return back()->with('message', 'Listing updated successfully..!');
     }
 
     //delete listing
-    public function destroy(Listing $listing){
+    public function destroy(Listing $listing)
+    {
+
+        //make sure logged in user is owner
+        if ($listing->user_id != auth()->id) {
+            abort(403, 'Unauthorized Action..!');
+        }
+
         $listing->delete();
-        return redirect('/')->with('message','Listing deleted successfylly..!');
+        return redirect('/')->with('message', 'Listing deleted successfylly..!');
     }
 
-
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
 }
